+++ R/tweetFetcher.R
@@ -1,6 +1,6 @@
 #' Fetch Tweets
 #'
-#' This function prepare http information for fetching process and working together sub functin
+#' This function prepare http information for fetching process and working together sub function
 #'
 #' @return list
 #' @export
@@ -26,38 +26,18 @@ tweetFetcher <- function() {
 # Twitter Recent Search Fetch
 tweetFetcher.fetch <- function(returnedInfo) {
   message("Analysis started !")
-  # https://stackoverflow.com/questions/12193779/how-to-write-trycatch-in-r
-  text = NULL
-  pb <- NULL
-  #scanned start 1 because also fetchedTweet variable have a fetched tweet
-  scanned <- 1
   fetchedTweet <- getTweet(returnedInfo)
-  isRate = fetchedTweet$remainingRateLimit
-  isRemain = if (is.numeric(isRate)) isRate else 1
-
-  if (is.numeric(isRate)){
-    # isRemain + 1 give correct percent information, because fetching one before start progress bar
-    pb = progress::progress_bar$new(format = "[:bar] :current/:total (:percent)", total = isRemain + 1)
+  textData <- NULL
+  while (length(fetchedTweet$text) > 0) {
+    if (!is.null(textData))
+      textData <- append(textData, fetchedTweet$text)
+    else
+      textData <- fetchedTweet$text
+    returnedInfo$params = fetchedTweet$nextParams
+    fetchedTweet <- getTweet(returnedInfo)
   }
-  if (!is.null(pb)){
-    # loop condition order is important, first of every time check fetchedTweet
-    while(!is.null(fetchedTweet) && scanned <= isRemain){
-      scanned = scanned + 1
-      text <- if (!is.null(text))
-        paste(text, fetchedTweet$text)
-      else
-        fetchedTweet$text
-      returnedInfo$params = fetchedTweet$nextParams
-      fetchedTweet <- getTweet(returnedInfo)
-      pb$tick()
-      Sys.sleep(1 / isRate)
-    }
-    # need completed %100 percent
-    pb$tick(isRemain + 1)
-    Sys.sleep(1 / isRate)
-    message(paste("\nScaning result :",scanned,"page scanned !"))
-    return(text)
-  }else{
-      return(NULL)
+  if (!is.null(fetchedTweet$news)){
+    message(news)
   }
+  return(textData)
 }